requirements.txt
# Core dependencies for Automatic LinkedIn Poster

# HTTP requests for LinkedIn API
requests>=2.31.0

# Google Sheets API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# LangChain and LangGraph for LLM workflow
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.20
langchain-core>=0.1.0

# Pydantic for data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
python-dotenv>=1.0.0


# Date/time handling for scheduling
python-dateutil>=2.8.2

# For background process
schedule>=1.2.0

# For Gradio setup interface and web UI
gradio>=4.0.0
pandas>=1.5.0

# Additional dependencies for web interface
uvicorn>=0.20.0
fastapi>=0.100.0
